model/requirements.txt
subprocess
sys
einops
tensorboard
torchvision
numpy
torch
Pillow
matplotlib
pytesseract

importlib-metadata
pathlib
typing
